requirements.txt
requests
python-dotenv
radon 
pylint
lizard
flask 
flask-cors
